requirements.txt
django==1.8.17
mysql-python
pytz
django-tinymce
django-widget-tweaks
django-markdownx